requirements.txt
ultralytics>=8.3.0
opencv-python-headless>=4.8.0.76
streamlit>=1.28.0
PyYAML>=6.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
